requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
spacy==3.7.2
nltk==3.8.1
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.25.0
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.10.2
openai==1.3.0
whisper
